+++ inaturalist_extractor/resources/styles/research_style.qml
@@ -1,0 +1,812 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis simplifyLocal="1" hasScaleBasedVisibilityFlag="0" simplifyMaxScale="1" version="3.42.0-Münster" minScale="100000000" simplifyAlgorithm="0" maxScale="0" autoRefreshMode="Disabled" autoRefreshTime="0" symbologyReferenceScale="-1" simplifyDrawingTol="1" labelsEnabled="0" simplifyDrawingHints="0" styleCategories="AllStyleCategories" readOnly="0">
+  <flags>
+    <Identifiable>1</Identifiable>
+    <Removable>1</Removable>
+    <Searchable>1</Searchable>
+    <Private>0</Private>
+  </flags>
+  <temporal enabled="0" mode="0" durationField="fid" startExpression="" fixedDuration="0" limitMode="0" durationUnit="min" endField="" accumulate="0" startField="" endExpression="">
+    <fixedRange>
+      <start></start>
+      <end></end>
+    </fixedRange>
+  </temporal>
+  <elevation respectLayerSymbol="1" showMarkerSymbolInSurfacePlots="0" clamping="Terrain" extrusionEnabled="0" zscale="1" binding="Centroid" extrusion="0" type="IndividualFeatures" symbology="Line" zoffset="0" customToleranceEnabled="0">
+    <data-defined-properties>
+      <Option type="Map">
+        <Option name="name" type="QString" value=""/>
+        <Option name="properties"/>
+        <Option name="type" type="QString" value="collection"/>
+      </Option>
+    </data-defined-properties>
+    <profileLineSymbol>
+      <symbol force_rhr="0" name="" is_animated="0" frame_rate="10" alpha="1" clip_to_extent="1" type="line">
+        <data_defined_properties>
+          <Option type="Map">
+            <Option name="name" type="QString" value=""/>
+            <Option name="properties"/>
+            <Option name="type" type="QString" value="collection"/>
+          </Option>
+        </data_defined_properties>
+        <layer id="{553a92cc-4b89-44c9-b9fe-4ddbd228248e}" pass="0" enabled="1" class="SimpleLine" locked="0">
+          <Option type="Map">
+            <Option name="align_dash_pattern" type="QString" value="0"/>
+            <Option name="capstyle" type="QString" value="square"/>
+            <Option name="customdash" type="QString" value="5;2"/>
+            <Option name="customdash_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="customdash_unit" type="QString" value="MM"/>
+            <Option name="dash_pattern_offset" type="QString" value="0"/>
+            <Option name="dash_pattern_offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="dash_pattern_offset_unit" type="QString" value="MM"/>
+            <Option name="draw_inside_polygon" type="QString" value="0"/>
+            <Option name="joinstyle" type="QString" value="bevel"/>
+            <Option name="line_color" type="QString" value="196,60,57,255,rgb:0.7686274509803922,0.23529411764705882,0.22352941176470589,1"/>
+            <Option name="line_style" type="QString" value="solid"/>
+            <Option name="line_width" type="QString" value="0.6"/>
+            <Option name="line_width_unit" type="QString" value="MM"/>
+            <Option name="offset" type="QString" value="0"/>
+            <Option name="offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="offset_unit" type="QString" value="MM"/>
+            <Option name="ring_filter" type="QString" value="0"/>
+            <Option name="trim_distance_end" type="QString" value="0"/>
+            <Option name="trim_distance_end_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="trim_distance_end_unit" type="QString" value="MM"/>
+            <Option name="trim_distance_start" type="QString" value="0"/>
+            <Option name="trim_distance_start_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="trim_distance_start_unit" type="QString" value="MM"/>
+            <Option name="tweak_dash_pattern_on_corners" type="QString" value="0"/>
+            <Option name="use_custom_dash" type="QString" value="0"/>
+            <Option name="width_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+          </Option>
+          <data_defined_properties>
+            <Option type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+          </data_defined_properties>
+        </layer>
+      </symbol>
+    </profileLineSymbol>
+    <profileFillSymbol>
+      <symbol force_rhr="0" name="" is_animated="0" frame_rate="10" alpha="1" clip_to_extent="1" type="fill">
+        <data_defined_properties>
+          <Option type="Map">
+            <Option name="name" type="QString" value=""/>
+            <Option name="properties"/>
+            <Option name="type" type="QString" value="collection"/>
+          </Option>
+        </data_defined_properties>
+        <layer id="{e8e42e95-5bca-44a2-893a-e15945fedb70}" pass="0" enabled="1" class="SimpleFill" locked="0">
+          <Option type="Map">
+            <Option name="border_width_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="color" type="QString" value="196,60,57,255,rgb:0.7686274509803922,0.23529411764705882,0.22352941176470589,1"/>
+            <Option name="joinstyle" type="QString" value="bevel"/>
+            <Option name="offset" type="QString" value="0,0"/>
+            <Option name="offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="offset_unit" type="QString" value="MM"/>
+            <Option name="outline_color" type="QString" value="140,43,41,255,rgb:0.5490196078431373,0.16803234912642101,0.15967040512703135,1"/>
+            <Option name="outline_style" type="QString" value="solid"/>
+            <Option name="outline_width" type="QString" value="0.2"/>
+            <Option name="outline_width_unit" type="QString" value="MM"/>
+            <Option name="style" type="QString" value="solid"/>
+          </Option>
+          <data_defined_properties>
+            <Option type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+          </data_defined_properties>
+        </layer>
+      </symbol>
+    </profileFillSymbol>
+    <profileMarkerSymbol>
+      <symbol force_rhr="0" name="" is_animated="0" frame_rate="10" alpha="1" clip_to_extent="1" type="marker">
+        <data_defined_properties>
+          <Option type="Map">
+            <Option name="name" type="QString" value=""/>
+            <Option name="properties"/>
+            <Option name="type" type="QString" value="collection"/>
+          </Option>
+        </data_defined_properties>
+        <layer id="{e9267ee1-cc0b-4746-8adb-fb4e8639daac}" pass="0" enabled="1" class="SimpleMarker" locked="0">
+          <Option type="Map">
+            <Option name="angle" type="QString" value="0"/>
+            <Option name="cap_style" type="QString" value="square"/>
+            <Option name="color" type="QString" value="196,60,57,255,rgb:0.7686274509803922,0.23529411764705882,0.22352941176470589,1"/>
+            <Option name="horizontal_anchor_point" type="QString" value="1"/>
+            <Option name="joinstyle" type="QString" value="bevel"/>
+            <Option name="name" type="QString" value="diamond"/>
+            <Option name="offset" type="QString" value="0,0"/>
+            <Option name="offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="offset_unit" type="QString" value="MM"/>
+            <Option name="outline_color" type="QString" value="140,43,41,255,rgb:0.5490196078431373,0.16803234912642101,0.15967040512703135,1"/>
+            <Option name="outline_style" type="QString" value="solid"/>
+            <Option name="outline_width" type="QString" value="0.2"/>
+            <Option name="outline_width_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="outline_width_unit" type="QString" value="MM"/>
+            <Option name="scale_method" type="QString" value="diameter"/>
+            <Option name="size" type="QString" value="3"/>
+            <Option name="size_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="size_unit" type="QString" value="MM"/>
+            <Option name="vertical_anchor_point" type="QString" value="1"/>
+          </Option>
+          <data_defined_properties>
+            <Option type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+          </data_defined_properties>
+        </layer>
+      </symbol>
+    </profileMarkerSymbol>
+  </elevation>
+  <renderer-v2 referencescale="-1" forceraster="0" attr="quality" type="categorizedSymbol" symbollevels="0" enableorderby="0">
+    <categories>
+      <category symbol="0" label="Needs id" render="true" uuid="{541a2d54-3cff-4825-9178-309523c75223}" type="string" value="needs_id"/>
+      <category symbol="1" label="Research" render="true" uuid="{02693230-f722-429c-9f6f-72b1afce0dff}" type="string" value="research"/>
+    </categories>
+    <symbols>
+      <symbol force_rhr="0" name="0" is_animated="0" frame_rate="10" alpha="1" clip_to_extent="1" type="marker">
+        <data_defined_properties>
+          <Option type="Map">
+            <Option name="name" type="QString" value=""/>
+            <Option name="properties"/>
+            <Option name="type" type="QString" value="collection"/>
+          </Option>
+        </data_defined_properties>
+        <layer id="{0fa66569-2e0a-460a-8846-5f6371116b53}" pass="0" enabled="1" class="SimpleMarker" locked="0">
+          <Option type="Map">
+            <Option name="angle" type="QString" value="0"/>
+            <Option name="cap_style" type="QString" value="square"/>
+            <Option name="color" type="QString" value="241,154,48,255,hsv:0.09166666666666666,0.80134279392690932,0.94558632791638053,1"/>
+            <Option name="horizontal_anchor_point" type="QString" value="1"/>
+            <Option name="joinstyle" type="QString" value="bevel"/>
+            <Option name="name" type="QString" value="circle"/>
+            <Option name="offset" type="QString" value="0,0"/>
+            <Option name="offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="offset_unit" type="QString" value="MM"/>
+            <Option name="outline_color" type="QString" value="255,255,255,255,hsv:0,0,1,1"/>
+            <Option name="outline_style" type="QString" value="solid"/>
+            <Option name="outline_width" type="QString" value="0.2"/>
+            <Option name="outline_width_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="outline_width_unit" type="QString" value="MM"/>
+            <Option name="scale_method" type="QString" value="diameter"/>
+            <Option name="size" type="QString" value="2.6"/>
+            <Option name="size_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="size_unit" type="QString" value="MM"/>
+            <Option name="vertical_anchor_point" type="QString" value="1"/>
+          </Option>
+          <data_defined_properties>
+            <Option type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+          </data_defined_properties>
+        </layer>
+      </symbol>
+      <symbol force_rhr="0" name="1" is_animated="0" frame_rate="10" alpha="1" clip_to_extent="1" type="marker">
+        <data_defined_properties>
+          <Option type="Map">
+            <Option name="name" type="QString" value=""/>
+            <Option name="properties"/>
+            <Option name="type" type="QString" value="collection"/>
+          </Option>
+        </data_defined_properties>
+        <layer id="{0fa66569-2e0a-460a-8846-5f6371116b53}" pass="0" enabled="1" class="SimpleMarker" locked="0">
+          <Option type="Map">
+            <Option name="angle" type="QString" value="0"/>
+            <Option name="cap_style" type="QString" value="square"/>
+            <Option name="color" type="QString" value="9,192,77,255,hsv:0.3953888888888889,0.95274280918593113,0.75460440985732813,1"/>
+            <Option name="horizontal_anchor_point" type="QString" value="1"/>
+            <Option name="joinstyle" type="QString" value="bevel"/>
+            <Option name="name" type="QString" value="circle"/>
+            <Option name="offset" type="QString" value="0,0"/>
+            <Option name="offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="offset_unit" type="QString" value="MM"/>
+            <Option name="outline_color" type="QString" value="255,255,255,255,hsv:0,0,1,1"/>
+            <Option name="outline_style" type="QString" value="solid"/>
+            <Option name="outline_width" type="QString" value="0.2"/>
+            <Option name="outline_width_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="outline_width_unit" type="QString" value="MM"/>
+            <Option name="scale_method" type="QString" value="diameter"/>
+            <Option name="size" type="QString" value="2.6"/>
+            <Option name="size_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="size_unit" type="QString" value="MM"/>
+            <Option name="vertical_anchor_point" type="QString" value="1"/>
+          </Option>
+          <data_defined_properties>
+            <Option type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+          </data_defined_properties>
+        </layer>
+      </symbol>
+    </symbols>
+    <source-symbol>
+      <symbol force_rhr="0" name="0" is_animated="0" frame_rate="10" alpha="1" clip_to_extent="1" type="marker">
+        <data_defined_properties>
+          <Option type="Map">
+            <Option name="name" type="QString" value=""/>
+            <Option name="properties"/>
+            <Option name="type" type="QString" value="collection"/>
+          </Option>
+        </data_defined_properties>
+        <layer id="{0fa66569-2e0a-460a-8846-5f6371116b53}" pass="0" enabled="1" class="SimpleMarker" locked="0">
+          <Option type="Map">
+            <Option name="angle" type="QString" value="0"/>
+            <Option name="cap_style" type="QString" value="square"/>
+            <Option name="color" type="QString" value="190,178,151,255,rgb:0.74509803921568629,0.69803921568627447,0.59215686274509804,1"/>
+            <Option name="horizontal_anchor_point" type="QString" value="1"/>
+            <Option name="joinstyle" type="QString" value="bevel"/>
+            <Option name="name" type="QString" value="circle"/>
+            <Option name="offset" type="QString" value="0,0"/>
+            <Option name="offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="offset_unit" type="QString" value="MM"/>
+            <Option name="outline_color" type="QString" value="255,255,255,255,hsv:0,0,1,1"/>
+            <Option name="outline_style" type="QString" value="solid"/>
+            <Option name="outline_width" type="QString" value="0.2"/>
+            <Option name="outline_width_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="outline_width_unit" type="QString" value="MM"/>
+            <Option name="scale_method" type="QString" value="diameter"/>
+            <Option name="size" type="QString" value="2.6"/>
+            <Option name="size_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="size_unit" type="QString" value="MM"/>
+            <Option name="vertical_anchor_point" type="QString" value="1"/>
+          </Option>
+          <data_defined_properties>
+            <Option type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+          </data_defined_properties>
+        </layer>
+      </symbol>
+    </source-symbol>
+    <rotation/>
+    <sizescale/>
+    <data-defined-properties>
+      <Option type="Map">
+        <Option name="name" type="QString" value=""/>
+        <Option name="properties"/>
+        <Option name="type" type="QString" value="collection"/>
+      </Option>
+    </data-defined-properties>
+  </renderer-v2>
+  <selection mode="Default">
+    <selectionColor invalid="1"/>
+    <selectionSymbol>
+      <symbol force_rhr="0" name="" is_animated="0" frame_rate="10" alpha="1" clip_to_extent="1" type="marker">
+        <data_defined_properties>
+          <Option type="Map">
+            <Option name="name" type="QString" value=""/>
+            <Option name="properties"/>
+            <Option name="type" type="QString" value="collection"/>
+          </Option>
+        </data_defined_properties>
+        <layer id="{9c589d16-0297-4909-873c-71d5f32b350c}" pass="0" enabled="1" class="SimpleMarker" locked="0">
+          <Option type="Map">
+            <Option name="angle" type="QString" value="0"/>
+            <Option name="cap_style" type="QString" value="square"/>
+            <Option name="color" type="QString" value="255,0,0,255,rgb:1,0,0,1"/>
+            <Option name="horizontal_anchor_point" type="QString" value="1"/>
+            <Option name="joinstyle" type="QString" value="bevel"/>
+            <Option name="name" type="QString" value="circle"/>
+            <Option name="offset" type="QString" value="0,0"/>
+            <Option name="offset_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="offset_unit" type="QString" value="MM"/>
+            <Option name="outline_color" type="QString" value="35,35,35,255,rgb:0.13725490196078433,0.13725490196078433,0.13725490196078433,1"/>
+            <Option name="outline_style" type="QString" value="solid"/>
+            <Option name="outline_width" type="QString" value="0"/>
+            <Option name="outline_width_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="outline_width_unit" type="QString" value="MM"/>
+            <Option name="scale_method" type="QString" value="diameter"/>
+            <Option name="size" type="QString" value="2"/>
+            <Option name="size_map_unit_scale" type="QString" value="3x:0,0,0,0,0,0"/>
+            <Option name="size_unit" type="QString" value="MM"/>
+            <Option name="vertical_anchor_point" type="QString" value="1"/>
+          </Option>
+          <data_defined_properties>
+            <Option type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+          </data_defined_properties>
+        </layer>
+      </symbol>
+    </selectionSymbol>
+  </selection>
+  <customproperties>
+    <Option type="Map">
+      <Option name="embeddedWidgets/count" type="int" value="0"/>
+      <Option name="variableNames"/>
+      <Option name="variableValues"/>
+    </Option>
+  </customproperties>
+  <blendMode>0</blendMode>
+  <featureBlendMode>0</featureBlendMode>
+  <layerOpacity>1</layerOpacity>
+  <geometryOptions removeDuplicateNodes="0" geometryPrecision="0">
+    <activeChecks/>
+    <checkConfiguration/>
+  </geometryOptions>
+  <legend showLabelLegend="0" type="default-vector"/>
+  <referencedLayers/>
+  <fieldConfiguration>
+    <field name="fid" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="id" configurationFlags="NoFlag">
+      <editWidget type="Range">
+        <config>
+          <Option type="Map">
+            <Option name="AllowNull" type="bool" value="true"/>
+            <Option name="Max" type="int" value="2147483647"/>
+            <Option name="Min" type="int" value="-2147483648"/>
+            <Option name="Precision" type="int" value="0"/>
+            <Option name="Step" type="int" value="1"/>
+            <Option name="Style" type="QString" value="SpinBox"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="iconic_taxon_name" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="taxon_id" configurationFlags="NoFlag">
+      <editWidget type="Range">
+        <config>
+          <Option type="Map">
+            <Option name="AllowNull" type="bool" value="true"/>
+            <Option name="Max" type="int" value="2147483647"/>
+            <Option name="Min" type="int" value="-2147483648"/>
+            <Option name="Precision" type="int" value="0"/>
+            <Option name="Step" type="int" value="1"/>
+            <Option name="Style" type="QString" value="SpinBox"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="rank" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="name" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="obs_login" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="obs_name" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="date" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="description" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="quality" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="geoprivacy" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="precision" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="url" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="taxon_url" configurationFlags="NoFlag">
+      <editWidget type="TextEdit">
+        <config>
+          <Option type="Map">
+            <Option name="IsMultiline" type="bool" value="false"/>
+            <Option name="UseHtml" type="bool" value="false"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="picture1" configurationFlags="NoFlag">
+      <editWidget type="ExternalResource">
+        <config>
+          <Option type="Map">
+            <Option name="DocumentViewer" type="int" value="1"/>
+            <Option name="DocumentViewerHeight" type="int" value="0"/>
+            <Option name="DocumentViewerWidth" type="int" value="0"/>
+            <Option name="FileWidget" type="bool" value="false"/>
+            <Option name="FileWidgetButton" type="bool" value="true"/>
+            <Option name="FileWidgetFilter" type="QString" value=""/>
+            <Option name="PropertyCollection" type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+            <Option name="RelativeStorage" type="int" value="0"/>
+            <Option name="StorageAuthConfigId" type="QString" value=""/>
+            <Option name="StorageMode" type="int" value="0"/>
+            <Option name="StorageType" type="QString" value="AWSS3"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="picture2" configurationFlags="NoFlag">
+      <editWidget type="ExternalResource">
+        <config>
+          <Option type="Map">
+            <Option name="DocumentViewer" type="int" value="1"/>
+            <Option name="DocumentViewerHeight" type="int" value="0"/>
+            <Option name="DocumentViewerWidth" type="int" value="0"/>
+            <Option name="FileWidget" type="bool" value="false"/>
+            <Option name="FileWidgetButton" type="bool" value="true"/>
+            <Option name="FileWidgetFilter" type="QString" value=""/>
+            <Option name="PropertyCollection" type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+            <Option name="RelativeStorage" type="int" value="0"/>
+            <Option name="StorageAuthConfigId" type="QString" value=""/>
+            <Option name="StorageMode" type="int" value="0"/>
+            <Option name="StorageType" type="QString" value="AWSS3"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+    <field name="picture3" configurationFlags="NoFlag">
+      <editWidget type="ExternalResource">
+        <config>
+          <Option type="Map">
+            <Option name="DocumentViewer" type="int" value="1"/>
+            <Option name="DocumentViewerHeight" type="int" value="0"/>
+            <Option name="DocumentViewerWidth" type="int" value="0"/>
+            <Option name="FileWidget" type="bool" value="false"/>
+            <Option name="FileWidgetButton" type="bool" value="true"/>
+            <Option name="FileWidgetFilter" type="QString" value=""/>
+            <Option name="PropertyCollection" type="Map">
+              <Option name="name" type="QString" value=""/>
+              <Option name="properties"/>
+              <Option name="type" type="QString" value="collection"/>
+            </Option>
+            <Option name="RelativeStorage" type="int" value="0"/>
+            <Option name="StorageAuthConfigId" type="QString" value=""/>
+            <Option name="StorageMode" type="int" value="0"/>
+            <Option name="StorageType" type="QString" value="AWSS3"/>
+          </Option>
+        </config>
+      </editWidget>
+    </field>
+  </fieldConfiguration>
+  <aliases>
+    <alias name="" index="0" field="fid"/>
+    <alias name="Id" index="1" field="id"/>
+    <alias name="Taxon name" index="2" field="iconic_taxon_name"/>
+    <alias name="Taxon Id" index="3" field="taxon_id"/>
+    <alias name="Rank" index="4" field="rank"/>
+    <alias name="Name" index="5" field="name"/>
+    <alias name="Observator login" index="6" field="obs_login"/>
+    <alias name="Observator name" index="7" field="obs_name"/>
+    <alias name="Date" index="8" field="date"/>
+    <alias name="Description" index="9" field="description"/>
+    <alias name="Quality" index="10" field="quality"/>
+    <alias name="Geoprivacy" index="11" field="geoprivacy"/>
+    <alias name="Precision" index="12" field="precision"/>
+    <alias name="Observation url" index="13" field="url"/>
+    <alias name="Taxon url" index="14" field="taxon_url"/>
+    <alias name="Picture 1" index="15" field="picture1"/>
+    <alias name="Picture 2" index="16" field="picture2"/>
+    <alias name="Picture 3" index="17" field="picture3"/>
+  </aliases>
+  <splitPolicies>
+    <policy policy="DefaultValue" field="fid"/>
+    <policy policy="DefaultValue" field="id"/>
+    <policy policy="DefaultValue" field="iconic_taxon_name"/>
+    <policy policy="DefaultValue" field="taxon_id"/>
+    <policy policy="DefaultValue" field="rank"/>
+    <policy policy="DefaultValue" field="name"/>
+    <policy policy="DefaultValue" field="obs_login"/>
+    <policy policy="DefaultValue" field="obs_name"/>
+    <policy policy="DefaultValue" field="date"/>
+    <policy policy="DefaultValue" field="description"/>
+    <policy policy="DefaultValue" field="quality"/>
+    <policy policy="DefaultValue" field="geoprivacy"/>
+    <policy policy="DefaultValue" field="precision"/>
+    <policy policy="DefaultValue" field="url"/>
+    <policy policy="DefaultValue" field="taxon_url"/>
+    <policy policy="DefaultValue" field="picture1"/>
+    <policy policy="DefaultValue" field="picture2"/>
+    <policy policy="DefaultValue" field="picture3"/>
+  </splitPolicies>
+  <duplicatePolicies>
+    <policy policy="Duplicate" field="fid"/>
+    <policy policy="Duplicate" field="id"/>
+    <policy policy="Duplicate" field="iconic_taxon_name"/>
+    <policy policy="Duplicate" field="taxon_id"/>
+    <policy policy="Duplicate" field="rank"/>
+    <policy policy="Duplicate" field="name"/>
+    <policy policy="Duplicate" field="obs_login"/>
+    <policy policy="Duplicate" field="obs_name"/>
+    <policy policy="Duplicate" field="date"/>
+    <policy policy="Duplicate" field="description"/>
+    <policy policy="Duplicate" field="quality"/>
+    <policy policy="Duplicate" field="geoprivacy"/>
+    <policy policy="Duplicate" field="precision"/>
+    <policy policy="Duplicate" field="url"/>
+    <policy policy="Duplicate" field="taxon_url"/>
+    <policy policy="Duplicate" field="picture1"/>
+    <policy policy="Duplicate" field="picture2"/>
+    <policy policy="Duplicate" field="picture3"/>
+  </duplicatePolicies>
+  <defaults>
+    <default expression="" field="fid" applyOnUpdate="0"/>
+    <default expression="" field="id" applyOnUpdate="0"/>
+    <default expression="" field="iconic_taxon_name" applyOnUpdate="0"/>
+    <default expression="" field="taxon_id" applyOnUpdate="0"/>
+    <default expression="" field="rank" applyOnUpdate="0"/>
+    <default expression="" field="name" applyOnUpdate="0"/>
+    <default expression="" field="obs_login" applyOnUpdate="0"/>
+    <default expression="" field="obs_name" applyOnUpdate="0"/>
+    <default expression="" field="date" applyOnUpdate="0"/>
+    <default expression="" field="description" applyOnUpdate="0"/>
+    <default expression="" field="quality" applyOnUpdate="0"/>
+    <default expression="" field="geoprivacy" applyOnUpdate="0"/>
+    <default expression="" field="precision" applyOnUpdate="0"/>
+    <default expression="" field="url" applyOnUpdate="0"/>
+    <default expression="" field="taxon_url" applyOnUpdate="0"/>
+    <default expression="" field="picture1" applyOnUpdate="0"/>
+    <default expression="" field="picture2" applyOnUpdate="0"/>
+    <default expression="" field="picture3" applyOnUpdate="0"/>
+  </defaults>
+  <constraints>
+    <constraint unique_strength="1" exp_strength="0" notnull_strength="1" field="fid" constraints="3"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="id" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="iconic_taxon_name" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="taxon_id" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="rank" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="name" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="obs_login" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="obs_name" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="date" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="description" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="quality" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="geoprivacy" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="precision" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="url" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="taxon_url" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="picture1" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="picture2" constraints="0"/>
+    <constraint unique_strength="0" exp_strength="0" notnull_strength="0" field="picture3" constraints="0"/>
+  </constraints>
+  <constraintExpressions>
+    <constraint exp="" desc="" field="fid"/>
+    <constraint exp="" desc="" field="id"/>
+    <constraint exp="" desc="" field="iconic_taxon_name"/>
+    <constraint exp="" desc="" field="taxon_id"/>
+    <constraint exp="" desc="" field="rank"/>
+    <constraint exp="" desc="" field="name"/>
+    <constraint exp="" desc="" field="obs_login"/>
+    <constraint exp="" desc="" field="obs_name"/>
+    <constraint exp="" desc="" field="date"/>
+    <constraint exp="" desc="" field="description"/>
+    <constraint exp="" desc="" field="quality"/>
+    <constraint exp="" desc="" field="geoprivacy"/>
+    <constraint exp="" desc="" field="precision"/>
+    <constraint exp="" desc="" field="url"/>
+    <constraint exp="" desc="" field="taxon_url"/>
+    <constraint exp="" desc="" field="picture1"/>
+    <constraint exp="" desc="" field="picture2"/>
+    <constraint exp="" desc="" field="picture3"/>
+  </constraintExpressions>
+  <expressionfields/>
+  <attributeactions>
+    <defaultAction key="Canvas" value="{00000000-0000-0000-0000-000000000000}"/>
+  </attributeactions>
+  <attributetableconfig sortOrder="0" sortExpression="" actionWidgetStyle="dropDown">
+    <columns>
+      <column hidden="0" name="fid" width="-1" type="field"/>
+      <column hidden="0" name="id" width="-1" type="field"/>
+      <column hidden="0" name="iconic_taxon_name" width="-1" type="field"/>
+      <column hidden="0" name="taxon_id" width="-1" type="field"/>
+      <column hidden="0" name="rank" width="-1" type="field"/>
+      <column hidden="0" name="name" width="-1" type="field"/>
+      <column hidden="0" name="obs_login" width="-1" type="field"/>
+      <column hidden="0" name="obs_name" width="-1" type="field"/>
+      <column hidden="0" name="date" width="-1" type="field"/>
+      <column hidden="0" name="description" width="-1" type="field"/>
+      <column hidden="0" name="quality" width="-1" type="field"/>
+      <column hidden="0" name="geoprivacy" width="-1" type="field"/>
+      <column hidden="0" name="precision" width="-1" type="field"/>
+      <column hidden="0" name="url" width="-1" type="field"/>
+      <column hidden="0" name="taxon_url" width="-1" type="field"/>
+      <column hidden="0" name="picture1" width="-1" type="field"/>
+      <column hidden="0" name="picture2" width="-1" type="field"/>
+      <column hidden="0" name="picture3" width="-1" type="field"/>
+      <column hidden="1" width="-1" type="actions"/>
+    </columns>
+  </attributetableconfig>
+  <conditionalstyles>
+    <rowstyles/>
+    <fieldstyles/>
+  </conditionalstyles>
+  <storedexpressions/>
+  <editform tolerant="1"></editform>
+  <editforminit/>
+  <editforminitcodesource>0</editforminitcodesource>
+  <editforminitfilepath></editforminitfilepath>
+  <editforminitcode><![CDATA[# -*- coding: utf-8 -*-
+"""
+QGIS forms can have a Python function that is called when the form is
+opened.
+
+Use this function to add extra logic to your forms.
+
+Enter the name of the function in the "Python Init function"
+field.
+An example follows:
+"""
+from qgis.PyQt.QtWidgets import QWidget
+
+def my_form_open(dialog, layer, feature):
+    geom = feature.geometry()
+    control = dialog.findChild(QWidget, "MyLineEdit")
+]]></editforminitcode>
+  <featformsuppress>0</featformsuppress>
+  <editorlayout>generatedlayout</editorlayout>
+  <editable>
+    <field name="date" editable="1"/>
+    <field name="description" editable="1"/>
+    <field name="fid" editable="1"/>
+    <field name="geoprivacy" editable="1"/>
+    <field name="iconic_taxon_name" editable="1"/>
+    <field name="id" editable="1"/>
+    <field name="name" editable="1"/>
+    <field name="obs_login" editable="1"/>
+    <field name="obs_name" editable="1"/>
+    <field name="picture1" editable="1"/>
+    <field name="picture2" editable="1"/>
+    <field name="picture3" editable="1"/>
+    <field name="precision" editable="1"/>
+    <field name="quality" editable="1"/>
+    <field name="rank" editable="1"/>
+    <field name="taxon_id" editable="1"/>
+    <field name="taxon_url" editable="1"/>
+    <field name="url" editable="1"/>
+  </editable>
+  <labelOnTop>
+    <field name="date" labelOnTop="0"/>
+    <field name="description" labelOnTop="0"/>
+    <field name="fid" labelOnTop="0"/>
+    <field name="geoprivacy" labelOnTop="0"/>
+    <field name="iconic_taxon_name" labelOnTop="0"/>
+    <field name="id" labelOnTop="0"/>
+    <field name="name" labelOnTop="0"/>
+    <field name="obs_login" labelOnTop="0"/>
+    <field name="obs_name" labelOnTop="0"/>
+    <field name="picture1" labelOnTop="0"/>
+    <field name="picture2" labelOnTop="0"/>
+    <field name="picture3" labelOnTop="0"/>
+    <field name="precision" labelOnTop="0"/>
+    <field name="quality" labelOnTop="0"/>
+    <field name="rank" labelOnTop="0"/>
+    <field name="taxon_id" labelOnTop="0"/>
+    <field name="taxon_url" labelOnTop="0"/>
+    <field name="url" labelOnTop="0"/>
+  </labelOnTop>
+  <reuseLastValue>
+    <field reuseLastValue="0" name="date"/>
+    <field reuseLastValue="0" name="description"/>
+    <field reuseLastValue="0" name="fid"/>
+    <field reuseLastValue="0" name="geoprivacy"/>
+    <field reuseLastValue="0" name="iconic_taxon_name"/>
+    <field reuseLastValue="0" name="id"/>
+    <field reuseLastValue="0" name="name"/>
+    <field reuseLastValue="0" name="obs_login"/>
+    <field reuseLastValue="0" name="obs_name"/>
+    <field reuseLastValue="0" name="picture1"/>
+    <field reuseLastValue="0" name="picture2"/>
+    <field reuseLastValue="0" name="picture3"/>
+    <field reuseLastValue="0" name="precision"/>
+    <field reuseLastValue="0" name="quality"/>
+    <field reuseLastValue="0" name="rank"/>
+    <field reuseLastValue="0" name="taxon_id"/>
+    <field reuseLastValue="0" name="taxon_url"/>
+    <field reuseLastValue="0" name="url"/>
+  </reuseLastValue>
+  <dataDefinedFieldProperties/>
+  <widgets/>
+  <previewExpression>"iconic_taxon_name"</previewExpression>
+  <mapTip enabled="1"></mapTip>
+  <layerGeometryType>0</layerGeometryType>
+</qgis>